Moved errors into error module
